Update README.md for HugeIcons-React
The README.md file has been updated with instructions on how to use the HugeIcons-React library, which includes over 3600 stroke icons. It also outlines the noteworthy features such as icon size and color customization, rotation, and animation, along with examples of usage for each feature. It welcomed contributions and provided licensing details.